+++ docs/release-notes.md
@@ -9,3 +9,4 @@
 - Add docker.pull task to pull all images used by the devstack
 - Add %-migrate and %-createsuperuser tasks to manage the database
 - Add git.clone task to the Makefile
+- Add provision task to the Makefile (mysql)